move addVhostOptions for vhost dropdown to layout.js
